refactor(cli): extract yaml batch executor
